move repl2 code into repl
