+++ repl/src/main/scala/quasar/repl/Paths.scala
@@ -15,7 +15,7 @@
  */
 
 package quasar
-package repl2
+package repl
 
 import slamdata.Predef._
